+++ src/level.c
@@ -1,8 +1,6 @@
 #include "global.h"
-#include "import.h"
 #include "gfx.h"
 #include "tile.h"
-#include "import.h"
 #include "level.h"
 #include "primitives.h"
 #include "segment.h"
