+++ tests/check_entity_manager.h
@@ -1,0 +1,4 @@
+#include <check.h>
+#include "../src/entity_manager.h"
+
+Suite *entity_manager_suite(void);
