--- tests/check_import.c
@@ -1,324 +1,0 @@
-#include <check.h>
-#include "../src/import.h"
-#include "../src/assets.h"
-#include "../src/sprites.h"
-#include "../src/files.h"
-#include "../src/level.h"
-#include "../src/gfx.h"
-#include "../src/light.h"
-
-
-START_TEST (IMP_read_level_check_negative)
-{
-    // GIVEN
-    level_t *level = NULL;
-    char *filename =    "./nonexisintg";
-
-    // WHEN
-    level = LVL_read_level(filename);
-
-    // THEN
-    ck_assert_ptr_null(level);
-}
-END_TEST
- 
-START_TEST (IMP_read_level_check_positive)
-{
-    // GIVEN
-    level_t *level = NULL;
-    char *filename = "./tests/testfiles/testlevel";
-
-    animations[ASSET_HERO_ANIMATION]         = ANIM_read_animation(FILEPATH_HERO_ANIMATION);
-    sprites[ASSET_SPRITE_HERO]               = GFX_read_texture(FILEPATH_SPRITE_HERO);
-
-    int expected_x_size = 40;
-    int expected_y_size = 40;
-    int expected_x_hero = 6 * TILE_WIDTH;
-    int expected_y_hero = 4 * TILE_HEIGHT;
-
-    // WHEN
-    level = LVL_read_level(filename);
-
-    // THEN
-    ck_assert_ptr_nonnull(level);
-
-    ck_assert_int_eq(level->size_x, expected_x_size);
-    ck_assert_int_eq(level->size_y, expected_y_size);
-
-    ck_assert_int_eq(LVL_hero_x(level), expected_x_hero);
-    ck_assert_int_eq(LVL_hero_y(level), expected_y_hero);
-}
-END_TEST
- 
-START_TEST (IMP_read_animation_check)
-{
-    // GIVEN
-    animation_sheet_t *sheet = NULL;
-    const char *data               = FILEPATH_HERO_ANIMATION;
-    const char *texture            = FILEPATH_SPRITE_HERO;
-
-    // WHEN
-    sheet = ANIM_read_animation(data);
-    sheet->texture = GFX_read_texture(texture);
-
-    // THEN
-    ck_assert_ptr_nonnull(sheet);
-    ck_assert_ptr_nonnull(sheet->texture);
-
-    // sheet
-    ck_assert_int_eq(sheet->n_animations, 4);
-
-    // single animation
-    ck_assert_int_eq(sheet->animations[0].len, 2);
-    ck_assert_int_eq(sheet->animations[1].len, 2);
-    ck_assert_int_eq(sheet->animations[2].len, 1);
-    ck_assert_int_eq(sheet->animations[3].len, 1);
-
-    // single frame
-    // first animation 
-    // first frame
-    ck_assert_int_eq(sheet->animations[0].frames[0].delay, 20);
-    ck_assert_int_eq(sheet->animations[0].frames[0].handle_x, 9);
-    ck_assert_int_eq(sheet->animations[0].frames[0].handle_y, 10);
-
-    ck_assert_int_eq(sheet->animations[0].frames[0].rect.x, 0);
-    ck_assert_int_eq(sheet->animations[0].frames[0].rect.y, 0);
-    ck_assert_int_eq(sheet->animations[0].frames[0].rect.w, 9);
-    ck_assert_int_eq(sheet->animations[0].frames[0].rect.h, 20);
-
-    ck_assert_int_eq(sheet->animations[0].frames[0].hit_boxes[0]. x, 0);
-    ck_assert_int_eq(sheet->animations[0].frames[0].hit_boxes[0]. y, 0);
-    ck_assert_int_eq(sheet->animations[0].frames[0].hit_boxes[0]. w, 9);
-    ck_assert_int_eq(sheet->animations[0].frames[0].hit_boxes[0]. h, 20);
-
-    //second frame
-    ck_assert_int_eq(sheet->animations[0].frames[1].delay, 20);
-    ck_assert_int_eq(sheet->animations[0].frames[1].handle_x, 9);
-    ck_assert_int_eq(sheet->animations[0].frames[1].handle_y, 10);
-
-    ck_assert_int_eq(sheet->animations[0].frames[1].rect.x, 9);
-    ck_assert_int_eq(sheet->animations[0].frames[1].rect.y, 0);
-    ck_assert_int_eq(sheet->animations[0].frames[1].rect.w, 9);
-    ck_assert_int_eq(sheet->animations[0].frames[1].rect.h, 20);
-
-    ck_assert_int_eq(sheet->animations[0].frames[1].hit_boxes[0]. x, 0);
-    ck_assert_int_eq(sheet->animations[0].frames[1].hit_boxes[0]. y, 0);
-    ck_assert_int_eq(sheet->animations[0].frames[1].hit_boxes[0]. w, 9);
-    ck_assert_int_eq(sheet->animations[0].frames[1].hit_boxes[0]. h, 20);
-
-    // second animation 
-    // first frame
-    ck_assert_int_eq(sheet->animations[1].frames[0].delay, 5);
-    ck_assert_int_eq(sheet->animations[1].frames[0].handle_x, 9);
-    ck_assert_int_eq(sheet->animations[1].frames[0].handle_y, 10);
-
-    ck_assert_int_eq(sheet->animations[1].frames[0].rect.x, 0);
-    ck_assert_int_eq(sheet->animations[1].frames[0].rect.y, 20);
-    ck_assert_int_eq(sheet->animations[1].frames[0].rect.w, 9);
-    ck_assert_int_eq(sheet->animations[1].frames[0].rect.h, 20);
-
-    ck_assert_int_eq(sheet->animations[1].frames[0].hit_boxes[0]. x, 0);
-    ck_assert_int_eq(sheet->animations[1].frames[0].hit_boxes[0]. y, 0);
-    ck_assert_int_eq(sheet->animations[1].frames[0].hit_boxes[0]. w, 9);
-    ck_assert_int_eq(sheet->animations[1].frames[0].hit_boxes[0]. h, 20);
-
-    // second frame
-    ck_assert_int_eq(sheet->animations[1].frames[1].delay, 5);
-    ck_assert_int_eq(sheet->animations[1].frames[1].handle_x, 9);
-    ck_assert_int_eq(sheet->animations[1].frames[1].handle_y, 10);
-
-    ck_assert_int_eq(sheet->animations[1].frames[1].rect.x, 9);
-    ck_assert_int_eq(sheet->animations[1].frames[1].rect.y, 20);
-    ck_assert_int_eq(sheet->animations[1].frames[1].rect.w, 9);
-    ck_assert_int_eq(sheet->animations[1].frames[1].rect.h, 20);
-
-    ck_assert_int_eq(sheet->animations[1].frames[1].hit_boxes[0]. x, 0);
-    ck_assert_int_eq(sheet->animations[1].frames[1].hit_boxes[0]. y, 0);
-    ck_assert_int_eq(sheet->animations[1].frames[1].hit_boxes[0]. w, 9);
-    ck_assert_int_eq(sheet->animations[1].frames[1].hit_boxes[0]. h, 20);
-
-    // third animation 
-    // first frame
-    ck_assert_int_eq(sheet->animations[2].frames[0].delay, 0);
-    ck_assert_int_eq(sheet->animations[2].frames[0].handle_x, 9);
-    ck_assert_int_eq(sheet->animations[2].frames[0].handle_y, 10);
-
-    ck_assert_int_eq(sheet->animations[2].frames[0].rect.x, 18);
-    ck_assert_int_eq(sheet->animations[2].frames[0].rect.y, 0);
-    ck_assert_int_eq(sheet->animations[2].frames[0].rect.w, 9);
-    ck_assert_int_eq(sheet->animations[2].frames[0].rect.h, 20);
-
-    ck_assert_int_eq(sheet->animations[2].frames[0].hit_boxes[0]. x, 0);
-    ck_assert_int_eq(sheet->animations[2].frames[0].hit_boxes[0]. y, 0);
-    ck_assert_int_eq(sheet->animations[2].frames[0].hit_boxes[0]. w, 9);
-    ck_assert_int_eq(sheet->animations[2].frames[0].hit_boxes[0]. h, 20);
-
-    // fourth animation 
-    // first frame
-    ck_assert_int_eq(sheet->animations[3].frames[0].delay, 0);
-    ck_assert_int_eq(sheet->animations[3].frames[0].handle_x, 9);
-    ck_assert_int_eq(sheet->animations[3].frames[0].handle_y, 10);
-
-    ck_assert_int_eq(sheet->animations[3].frames[0].rect.x, 18);
-    ck_assert_int_eq(sheet->animations[3].frames[0].rect.y, 20);
-    ck_assert_int_eq(sheet->animations[3].frames[0].rect.w, 9);
-    ck_assert_int_eq(sheet->animations[3].frames[0].rect.h, 20);
-
-    ck_assert_int_eq(sheet->animations[3].frames[0].hit_boxes[0]. x, 0);
-    ck_assert_int_eq(sheet->animations[3].frames[0].hit_boxes[0]. y, 0);
-    ck_assert_int_eq(sheet->animations[3].frames[0].hit_boxes[0]. w, 9);
-    ck_assert_int_eq(sheet->animations[3].frames[0].hit_boxes[0]. h, 20);
-}
-END_TEST
-
-START_TEST (IMP_read_wobble_check)
-{
-    // GIVEN
-    wobble_t *wobble = NULL;
-    char *filename   = "./assets/wobbles/walking.wbl";
-
-    // WHEN
-    wobble = LIG_read_wobble(filename);
-
-    // THEN
-    ck_assert_ptr_nonnull(wobble);
-    ck_assert_int_eq(wobble->len, 15);
-
-    ck_assert_float_eq(wobble->coefs[0], 0.012);
-    ck_assert_float_eq(wobble->coefs[1], 0.054);
-    ck_assert_float_eq(wobble->coefs[2], 0.082);
-    ck_assert_float_eq(wobble->coefs[3], 0.1);
-    ck_assert_float_eq(wobble->coefs[4], 0.082);
-    ck_assert_float_eq(wobble->coefs[5], 0.054);
-    ck_assert_float_eq(wobble->coefs[6], 0.012);
-    ck_assert_float_eq(wobble->coefs[7], 0.0);
-    ck_assert_float_eq(wobble->coefs[8], -0.012);
-    ck_assert_float_eq(wobble->coefs[9], -0.054);
-    ck_assert_float_eq(wobble->coefs[10], -0.082);
-    ck_assert_float_eq(wobble->coefs[11], -0.1);
-    ck_assert_float_eq(wobble->coefs[12], -0.082);
-    ck_assert_float_eq(wobble->coefs[13], -0.054);
-    ck_assert_float_eq(wobble->coefs[14], -0.012);
-}
-END_TEST
-
-START_TEST (IMP_read_lightsource_lighter_check)
-{
-    // GIVEN
-    lightsource_t *lightsource  = NULL;
-    char *filename              = "./assets/lightsources/lighter.lsrc";
-
-    // WHEN
-    lightsource = LIG_read_lightsource(filename);
-
-    // THEN
-    ck_assert_ptr_nonnull(lightsource);
-
-    ck_assert_float_eq(lightsource->width, PI / 7);
-    ck_assert_int_eq(lightsource->penetrating_power, 7);
-    ck_assert_int_eq(lightsource->n_poly, 3);
-
-    ck_assert_int_eq(lightsource->light_polygons[0].x, 0);
-    ck_assert_int_eq(lightsource->light_polygons[0].y, 0);
-    ck_assert_int_eq(lightsource->light_polygons[0].red, 255);
-    ck_assert_int_eq(lightsource->light_polygons[0].green, 251);
-    ck_assert_int_eq(lightsource->light_polygons[0].blue, 187);
-    ck_assert_int_eq(lightsource->light_polygons[0].light_power, 50);
-    ck_assert_int_eq(lightsource->light_polygons[0].width, 0);
-
-    ck_assert_int_eq(lightsource->light_polygons[1].x, 0);
-    ck_assert_int_eq(lightsource->light_polygons[1].y, 0);
-    ck_assert_int_eq(lightsource->light_polygons[1].red, 255);
-    ck_assert_int_eq(lightsource->light_polygons[1].green, 251);
-    ck_assert_int_eq(lightsource->light_polygons[1].blue, 187);
-    ck_assert_int_eq(lightsource->light_polygons[1].light_power, 30);
-    ck_assert_int_eq(lightsource->light_polygons[1].width, 36);
-
-    ck_assert_int_eq(lightsource->light_polygons[2].x, 0);
-    ck_assert_int_eq(lightsource->light_polygons[2].y, 0);
-    ck_assert_int_eq(lightsource->light_polygons[2].red, 255);
-    ck_assert_int_eq(lightsource->light_polygons[2].green, 251);
-    ck_assert_int_eq(lightsource->light_polygons[2].blue, 187);
-    ck_assert_int_eq(lightsource->light_polygons[2].light_power, 20);
-    ck_assert_int_eq(lightsource->light_polygons[2].width, 72);
-}
-END_TEST
-
-START_TEST (IMP_read_lightsource_lantern_check)
-{
-    // GIVEN
-    lightsource_t *lightsource  = NULL;
-    char *filename         = "./assets/lightsources/lantern.lsrc";
-
-    // WHEN
-    lightsource = LIG_read_lightsource(filename);
-
-    // THEN
-    ck_assert_ptr_nonnull(lightsource);
-
-    ck_assert_float_eq(lightsource->width, 0.0);
-    ck_assert_int_eq(lightsource->penetrating_power, 7);
-    ck_assert_int_eq(lightsource->n_poly, 5);
-
-    ck_assert_int_eq(lightsource->light_polygons[0].x, -10);
-    ck_assert_int_eq(lightsource->light_polygons[0].y, -10);
-    ck_assert_int_eq(lightsource->light_polygons[0].red, 255);
-    ck_assert_int_eq(lightsource->light_polygons[0].green, 251);
-    ck_assert_int_eq(lightsource->light_polygons[0].blue, 187);
-    ck_assert_int_eq(lightsource->light_polygons[0].light_power, 10);
-    ck_assert_int_eq(lightsource->light_polygons[0].width, 0);
-
-    ck_assert_int_eq(lightsource->light_polygons[1].x, 10);
-    ck_assert_int_eq(lightsource->light_polygons[1].y, -10);
-    ck_assert_int_eq(lightsource->light_polygons[1].red, 255);
-    ck_assert_int_eq(lightsource->light_polygons[1].green, 251);
-    ck_assert_int_eq(lightsource->light_polygons[1].blue, 187);
-    ck_assert_int_eq(lightsource->light_polygons[1].light_power, 10);
-    ck_assert_int_eq(lightsource->light_polygons[1].width, 0);
-
-    ck_assert_int_eq(lightsource->light_polygons[2].x, 10);
-    ck_assert_int_eq(lightsource->light_polygons[2].y, 10);
-    ck_assert_int_eq(lightsource->light_polygons[2].red, 255);
-    ck_assert_int_eq(lightsource->light_polygons[2].green, 251);
-    ck_assert_int_eq(lightsource->light_polygons[2].blue, 187);
-    ck_assert_int_eq(lightsource->light_polygons[2].light_power, 10);
-    ck_assert_int_eq(lightsource->light_polygons[2].width, 0);
-
-    ck_assert_int_eq(lightsource->light_polygons[3].x, -10);
-    ck_assert_int_eq(lightsource->light_polygons[3].y, 10);
-    ck_assert_int_eq(lightsource->light_polygons[3].red, 255);
-    ck_assert_int_eq(lightsource->light_polygons[3].green, 251);
-    ck_assert_int_eq(lightsource->light_polygons[3].blue, 187);
-    ck_assert_int_eq(lightsource->light_polygons[3].light_power, 10);
-    ck_assert_int_eq(lightsource->light_polygons[3].width, 0);
-
-    ck_assert_int_eq(lightsource->light_polygons[4].x, 0);
-    ck_assert_int_eq(lightsource->light_polygons[4].y, 0);
-    ck_assert_int_eq(lightsource->light_polygons[4].red, 255);
-    ck_assert_int_eq(lightsource->light_polygons[4].green, 251);
-    ck_assert_int_eq(lightsource->light_polygons[4].blue, 187);
-    ck_assert_int_eq(lightsource->light_polygons[4].light_power, 50);
-    ck_assert_int_eq(lightsource->light_polygons[4].width, 0);
-
-}
-END_TEST
-
-Suite *import_suite(void)
-{
-    Suite *s;
-    TCase *tc_core;
-
-    s = suite_create("import");
-    tc_core = tcase_create("Core");
-
-    tcase_add_test(tc_core, IMP_read_level_check_negative);
-    tcase_add_test(tc_core, IMP_read_level_check_positive);
-    tcase_add_test(tc_core, IMP_read_animation_check);
-    tcase_add_test(tc_core, IMP_read_wobble_check);
-    tcase_add_test(tc_core, IMP_read_lightsource_lighter_check);
-    tcase_add_test(tc_core, IMP_read_lightsource_lantern_check);
-
-    suite_add_tcase(s, tc_core);
-
-    return s;
-}
