--- tests/check_import.h
@@ -1,4 +1,0 @@
-#include <check.h>
-#include "../src/import.h"
-
-Suite* import_suite(void);
