+++ tests/check_level.c
@@ -180,76 +180,6 @@ START_TEST (LVL_fill_structure_check)
 }
 END_TEST
 
-START_TEST (LVL_entity_in_range_check)
-{
-    // GIVEN 
-    int size_x = 10; 
-    int size_y = 10;
-    int hero_x = 5;
-    int hero_y = 5;
-    
-    bool res    = false;
-    entity_t *e = NULL;
-
-    level_t* level = NULL;
-    level          = LVL_new();
-    level->size_y  = size_y;
-    level->size_x  = size_x;
-    
-    animations[ASSET_HERO_ANIMATION]         = ANIM_read_animation(FILEPATH_HERO_ANIMATION);
-    sprites[ASSET_SPRITE_HERO]               = GFX_read_texture(FILEPATH_SPRITE_HERO);
-    animations[ASSET_HERO_TEST]              = NULL;
-    sprites[ASSET_SPRITE_TEST]               = NULL;
-
-    // hero
-    LVL_add_entity(level, ENT_generate(hero_x, hero_y, ENTITY_HERO), 0);
-
-    e = NULL;
-    e = ENT_generate(24, 24, ENTITY_HERO);
-    res = LVL_entity_in_range(level, e, ENTITY_UPDATE_X_RANGE, ENTITY_UPDATE_Y_RANGE);
-    ck_assert_int_eq(res, 0);
-}
-
-START_TEST (LVL_apply_collision_check)
-{
-    // GIVEN 
-    int size_x = 10; 
-    int size_y = 10;
-    int entity_x = 5;
-    int entity_y = 5;
-
-    level_t* level = NULL;
-    level          = LVL_new();
-    level->size_y  = size_y;
-    level->size_x  = size_x;
-    
-    animations[ASSET_HERO_ANIMATION]         = ANIM_read_animation(FILEPATH_HERO_ANIMATION);
-    sprites[ASSET_SPRITE_HERO]               = GFX_read_texture(FILEPATH_SPRITE_HERO);
-
-    // hero
-    entity_t* e  = NULL;
-    e            = ENT_generate(entity_x, entity_y, ENTITY_HERO);
-    e->state     = WALKING;
-
-    // CASE 1
-    segment_t* obstacles = NULL;
-    SEG_push(&obstacles, 0, entity_y * TILE_HEIGHT + 5, 1000, entity_y * TILE_HEIGHT + 5);
-    level->obstacle_segments = obstacles;
-    LVL_apply_collision(level, e);
-    ck_assert_int_ne(e->y, entity_y*TILE_HEIGHT);
-
-    // CASE 2
-    obstacles    = NULL;
-    e            = ENT_generate(entity_x, entity_y, ENTITY_HERO);
-    e->state     = WALKING;
-    e->x_vel     = 2;
-    SEG_push(&obstacles, entity_x * TILE_WIDTH + 2, 0, entity_x * TILE_WIDTH + 2, 1000);
-    level->obstacle_segments = NULL;
-    level->obstacle_segments = obstacles;
-    LVL_apply_collision(level, e);
-    ck_assert_int_ne(e->x, entity_x*TILE_WIDTH);
-}
-
 Suite *level_suite(void)
 {
     Suite *s;
@@ -263,8 +193,6 @@ Suite *level_suite(void)
     tcase_add_test(tc_core, LVL_analyze_check);
     tcase_add_test(tc_core, LVL_fill_tiles_check);
     tcase_add_test(tc_core, LVL_fill_structure_check);
-    tcase_add_test(tc_core, LVL_entity_in_range_check);
-    tcase_add_test(tc_core, LVL_apply_collision_check);
 
     suite_add_tcase(s, tc_core);
