+++ tests/check_source.h
@@ -1,0 +1,4 @@
+#include <check.h>
+#include "../src/source.h"
+
+Suite *source_suite(void);
