+++ tests/check_sprites.h
@@ -1,0 +1,4 @@
+#include <check.h>
+#include "../src/sprites.h"
+
+Suite* sprites_suite(void);
